refactor ValidateOperation into a static part representing the operation and a dynamic part representing each validation instance corresponding to a particular table to be validated.
