+++ robot-core/src/test/resources/immune_exposures-result.txt
@@ -1,3 +1,4 @@
+Validating /home/mike/Knocean/robot/robot-core/target/test-classes/immune_exposures.csv ...
 At row: 17, column: 2: Cell is empty but rule: "is-required true" does not allow this.
 At row: 17, column: 4: Unable to parse rule "subclass-of hasMaterialBasisIn some (null)".
 At row: 17, column: 4: Validation failed for rule: "dengue hemorrhagic fever subclass-of hasMaterialBasisIn some (null)".
